[11519] Update Model Data Row Show Page UI
hamlize the show page

add a "detailed" option to the data show helper

add private status to the project header

this is in the data show page, but may as well fix
it in this commit set.

overhaul the design of the show page.

I am not super happy with this, but it will do for
now.